Linvis 1.0: Initial working design
Search Features:

Allows users to narrow the search results by release, and further specify the
author, commit id, or term in the preview of the commit.

Users can also specify merges by Linus that meet the requirement as well. All
results contain at least one commit or merge matching the search filters.

Search Results View:

The results view allows users to see all commits and merges that match their search
query. The results are presented in a table with the columns being the Log preview,
the author name, the date the commit or merge was authored, and the date the commit
or merge was committed. (these differ if a rebase occurs).

The table is fully paginated, allowing the user to limit the number of results they
see. The table can be re-ordered by any of the columns. Furthermore, an additional
search box is available allowing users to type text which will be matched against
any text within the table.

Commits View:

Clicking on any of the entries in the search results table will land the user at the
information view for that commit or merge. This view includes "breadcrumbs" for that
commit, showing the path that this commit will take before being merged into the
master fork of the kernel. This information is presented by CID, though it should be
updated to show a truncated form of the log preview. This view also includes the log
preview and four tabs.
The first tab is the message, it shows the entire commit log.
The second tab is the Files tab, showing what files were edited somewhere in the
commits within a merge, or in the commit. If the selected result is a commit, the
lines added and removed are simply the lines added and removed from the files in the
commit. If the result is a merge; however, the lines added and removed are simply
the sum of the lines added to the file in all of the sub-commits and sub-merges.
Likewise for the lines removed. It is possible that in one commit, a file may have a
line added, and then immediately have that line removed in another commit higher in
the tree. This will not result in the effect being canceled, this line being added
and removed will result in both the added lines being incremented and the removed
lines being incremented.
The third tab are the modules in the merge. We use a basic heuristic to gather the
module name; looking at the preview of the commit and extracting all text up to the
first colon. This appears to be a convention within the Linux kernel that is only
broken a few times and therefore provides fairly decent results. In a few cases this
is a filename, but for the most part is a good indicator of what is happening within
the merge. If the result is a commit, then there will only be one module name since
commits are leaf-nodes in the tree design; this module name will be the text as
described by the heuristic. If the result is a merge, there may be many commits
within that merge node containing the same module, in which case the module count is
increased, showing the sum of all commits with that module name.
The last tab is the tree-view. At this time, the tree-view is simply a nested list
design. The first entry is the current merge or commit, indentation shows how deep
within the tree a given commit or merge is. This is where additional work must be
done to perform full visualizations using a tree-based model of the Linux kernel
development.