Adding Reingold-Tilford Tree:
Re-implements the reigold tilford tree visualization.
The tree implementation is strictly svg, so vectorized output looks good. We were
using the css to assist with the tree before, so the vectorized output looked bad.
This is not the case anymore. The tree has been turned 90° so depth is going down.
I'm not certain that this is a good thing, the massive trees still don't fit on the
page very well, as is expected.