Dynamic Colour Range, Dynamic Tree Sizing
The old color range went from eggplant purple to periwinkle blue. The new color
range is simply shades of blue. The node with the most children will ALWAYS have the
darkest color, the nodes with fewer children will gradually become lighter. The leaf
nodes will always be white. The selected node will always be red.

The height and width of the tree is sized dynamically given the number of nodes,
height of the tree, and other various aspects of the tree.