Authorship: Provides information about commit authors
Table provides information on the number of lines added and removed by
an author within a merge. Also provides the total number of lines
modified in the merge.

Delta is the difference of the lines added from the lines removed, we
can see who is removing more lines versuse adding lines.

This is a raw sum and difference operation, we don't actually know how
many of these lines are simply being replaced.