Fixes Authorship Output -- BREAKING
This fixes the authorship information for merges and commits. The issue
was that only a single file was being counted for each author. The
change instead uses a list to list each item.

This change moves a bit of work from the server onto the client. Changes
may come in the future, moving things back into an object structure.