Add JS/CSS minification, Caching, JS/Css re-ordering
This combines and minimizes groups of javascript used in the page.
This minimizes the css for the spinner in the page. The bootstrap is
provided by the CDN.

Adds heavy caching to improve performance.

Re-orders JS/CSS to ensure that the page is downloaded and rendered
before unnecessary parts are downloaded.

Most of this is to improve the performance and get a better Google score.