Adding nested details table to author data
The nested table provides more information about what contributions the
author made to a given merge. It lists out the files, which commit the
file is attributed to, and how many lines were added and removed from a
file in a given commit. Furthermore, it provides a link directly to that
commit.