Initial Search Results update
Combining commits into groups based on which merge-tree they are part
of. The trees are ordered based on the normalized rank; the sum of the
ranks of each of the commits divided by the total number of commits.

The results look fine, but performance is terrible. Will need to work on
that. Needs server-side pagination, and better way of rendering each
table.